Added slicing by Oxz with overlapping.

+++ iosystem/src/main/java/Triangle.java
@@ -2,25 +2,25 @@
  * Created by Anton Klimansky on 02.03.2017.
  */
 public class Triangle {
-    private long p1;
-    private long p2;
-    private long p3;
+    private int p1;
+    private int p2;
+    private int p3;
 
-    public Triangle(long p1, long p2, long p3) {
+    public Triangle(int p1, int p2, int p3) {
         this.p1 = p1;
         this.p2 = p2;
         this.p3 = p3;
     }
 
-    public long getP1() {
+    public int getP1() {
         return p1;
     }
 
-    public long getP2() {
+    public int getP2() {
         return p2;
     }
 
-    public long getP3() {
+    public int getP3() {
         return p3;
     }
 }
